+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# unreleased
+
+- Add `StreamExt::switch` for dynamically switching over from one stream to the
+  next as soon as it's ready
+
 # 0.1.1
 
 - Add `StreamExt::batch_with` for flexible batching of the stream's items